Use int32 in Go interfaces that map to protobufs using int32 (#17068)
We use plain `int`, which can be 32 or 64 bit based on the system, in
some of the Go interfaces that map to protobuf using int32. Update the
Go interfaces to use int32 to prevent potential integer overflow.

These potential integer conversion overflows are flagged by the latest
golangci-lint. Once this PR is merged, we can upgrade our version of
golangci-lint, which will unblock us from using Go 1.23 in CI.

This will require some small updates in providers, mostly removing casts
`int32(...)`.

+++ pkg/resource/deploy/deploytest/provider_test.go
@@ -76,7 +76,7 @@ func TestProvider(t *testing.T) {
 			var called bool
 			prov := &Provider{
 				GetSchemaF: func(_ context.Context, req plugin.GetSchemaRequest) (plugin.GetSchemaResponse, error) {
-					assert.Equal(t, 1, req.Version)
+					assert.Equal(t, int32(1), req.Version)
 					assert.Equal(t, "expected-subpackage", req.SubpackageName)
 					assert.Equal(t, &expectedVersion, req.SubpackageVersion)
 					called = true
